cloudrun/requirements.txt
# Minimal build-time dependencies only
flask
google-cloud-storage

# Everything else installed at runtime
